Adit | fix some test for increase coverage

+++ survey/input.go
@@ -1,7 +1,7 @@
 package survey
 
 type CreateSurveyInput struct {
-	UserId int64 `json:"user_id"`
+	UserId uint `json:"user_id"`
 	SurveyCategory    string `json:"survey_category" binding:"required"`
 	SurveyTitle       string `json:"survey_title" binding:"required"`
 	SurveyDescription string `json:"survey_description" binding:"required"`
@@ -10,8 +10,8 @@ type CreateSurveyInput struct {
 }
 
 type QuestionInput struct {
-	SurveyId     int64 `json:"survey_id"`
-	UserId       int64 `json:"user_id"`
+	SurveyId     uint `json:"survey_id"`
+	UserId       uint `json:"user_id"`
 	SurveyQuestion string `json:"survey_question" binding:"required"`
 	QuestionType string `json:"question_type" binding:"required"`
 	OptionName   string `json:"option_name" binding:"required"`
